s:Rename to `ActiveToken`
Consistently use the singular for names of database columns and tables.
We also update the modules names to use the singular form.
