+++ lib/pbench/cli/server/user_management.py
@@ -3,7 +3,7 @@ import click
 from pbench import BadConfig
 from pbench.cli.server import config_setup, pass_cli_context
 from pbench.cli.server.options import common_options
-from pbench.server.database.models.users import Roles, User
+from pbench.server.database.models.user import Roles, User
 
 USER_LIST_ROW_FORMAT = "{0:15}\t{1:15}\t{2:15}\t{3:15}\t{4:20}"
 USER_LIST_HEADER_ROW = USER_LIST_ROW_FORMAT.format(
