+++ lib/pbench/server/api/resources/datasets_daterange.py
@@ -15,7 +15,7 @@ from pbench.server.api.resources import (
     Schema,
 )
 from pbench.server.database.database import Database
-from pbench.server.database.models.datasets import Dataset
+from pbench.server.database.models.dataset import Dataset
 
 
 class DatasetsDateRange(ApiBase):
