+++ lib/pbench/server/api/resources/query_apis/datasets_delete.py
@@ -15,7 +15,7 @@ from pbench.server.api.resources import (
 from pbench.server.api.resources.query_apis import ApiContext, ElasticBulkBase
 from pbench.server.cache_manager import CacheManager
 from pbench.server.database.models.audit import AuditType
-from pbench.server.database.models.datasets import Dataset, States
+from pbench.server.database.models.dataset import Dataset, States
 
 
 class DatasetsDelete(ElasticBulkBase):
