+++ lib/pbench/server/unpack_tarballs.py
@@ -5,7 +5,7 @@ import tempfile
 
 from pbench.server import PbenchServerConfig
 from pbench.server.cache_manager import CacheManager
-from pbench.server.database.models.datasets import Dataset, Metadata
+from pbench.server.database.models.dataset import Dataset, Metadata
 from pbench.server.report import Report
 from pbench.server.sync import Operation, Sync
