+++ lib/pbench/test/unit/server/query_apis/test_datasets_delete.py
@@ -7,7 +7,7 @@ import pytest
 
 from pbench.server import JSON, PbenchServerConfig
 from pbench.server.cache_manager import CacheManager
-from pbench.server.database.models.datasets import Dataset, DatasetNotFound
+from pbench.server.database.models.dataset import Dataset, DatasetNotFound
 from pbench.test.unit.server.headertypes import HeaderTypes
