+++ lib/pbench/test/unit/server/test_datasets_inventory.py
@@ -6,7 +6,7 @@ import requests
 import werkzeug.utils
 
 from pbench.server.cache_manager import CacheManager
-from pbench.server.database.models.datasets import Dataset, DatasetNotFound
+from pbench.server.database.models.dataset import Dataset, DatasetNotFound
 
 
 class TestDatasetsAccess:
